+++ backport/v19/core/class/commonhookactions.class.php
@@ -2,53 +2,45 @@
 
 namespace userNavHistory;
 
-/* Copyright (C) 2023 Laurent Destailleur  <eldy@users.sourceforge.net>
+/* Copyright (C) 2023 Laurent Destailleur <eldy@users.sourceforge.net>
  *
- * This program is free software; you can redistribute it and/or modify
+ * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 3 of the License, or
+ * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-/**
- *	\file       htdocs/core/class/commonhookactions.class.php
- *	\ingroup    core
- *	\brief      File of parent class of all other hook actions classes
- */
+if (file_exists(DOL_DOCUMENT_ROOT.'/core/class/commonhookactions.class.php')) {
+	require_once DOL_DOCUMENT_ROOT.'/core/class/commonhookactions.class.php';
 
-
-if (file_exists(DOL_DOCUMENT_ROOT . '/core/class/commonhookactions.class.php')) {
-	require_once DOL_DOCUMENT_ROOT . '/core/class/commonhookactions.class.php';
 	/**
-	 *	Parent class of all other hook actions classes
+	 * Compatibility parent using the native Dolibarr implementation.
 	 */
 	abstract class RetroCompatCommonHookActions extends \CommonHookActions
 	{
-
 	}
 } else {
 	/**
-	 *	Parent class of all other hook actions classes
+	 * Minimal compatibility parent for Dolibarr versions without CommonHookActions.
 	 */
 	abstract class RetroCompatCommonHookActions
 	{
 		/**
-		 * @var string	String of results.
+		 * @var string|null String displayed by executeHooks().
 		 */
 		public $resprints;
 
 		/**
-		 * @var array 	Array of results.
+		 * @var array<string, mixed> Hook results.
 		 */
 		public $results = array();
 	}
-
 }