+++ lib/usernavhistory.lib.php
@@ -40,6 +40,11 @@ function usernavhistoryAdminPrepareHead()
 	$head[$h][2] = 'settings';
 	$h++;
 
+	$head[$h][0] = dol_buildpath('/usernavhistory/admin/compatibility.php', 1);
+	$head[$h][1] = $langs->trans('Compatibility');
+	$head[$h][2] = 'compatibility';
+	$h++;
+
 	/*
 	$head[$h][0] = dol_buildpath("/usernavhistory/admin/myobject_extrafields.php", 1);
 	$head[$h][1] = $langs->trans("ExtraFields");